requirements.txt
Flask==2.0.2
Werkzeug==2.0.3  # Pin Werkzeug to a compatible version
flask-cors==3.0.10
numpy==1.22.0
librosa==0.9.1
torch==1.13.1
transformers==4.26.0
openai-whisper==20230314
soundfile==0.10.3.post1
opencv-python==4.8.0.74
scikit-image==0.21.0